Add filterQuery validation to inventory alert
