+++ x-pack/plugins/infra/server/lib/alerting/common/utils.ts
@@ -1,0 +1,26 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+import { isEmpty } from 'lodash';
+import { schema } from '@kbn/config-schema';
+
+export const oneOfLiterals = (arrayOfLiterals: Readonly<string[]>) =>
+  schema.string({
+    validate: value =>
+      arrayOfLiterals.includes(value) ? undefined : `must be one of ${arrayOfLiterals.join(' | ')}`,
+  });
+
+export const validateIsStringElasticsearchJSONFilter = (value: string) => {
+  const errorMessage = 'filterQuery must be a valid Elasticsearch filter expressed in JSON';
+  try {
+    const parsedValue = JSON.parse(value);
+    if (!isEmpty(parsedValue.bool)) {
+      return undefined;
+    }
+    return errorMessage;
+  } catch (e) {
+    return errorMessage;
+  }
+};
